Implemented encryption for passwords and made account fields private
A few spelling mistakes were corrected.

Class fields in account were changed from public to private and all
needed changes were made.

Implemented the following methods:
getPassword()
hashPassword()
toHexString()

The encryption that I implemented is MD5, a relatively secure
algorithm. This can be changed in hashPassword() if it is not good
enough.